Types of Constructors
1.Default Constructors:
 Default constructor is the constructor which doesn’t take any argument. It has no parameters.
2.Parameterized Constructors:
It is possible to pass arguments to constructors. Typically, these arguments help initialize an object when it is created. To create a parameterized constructor, simply add parameters to it the way you would to any other function. When you define the constructor’s body, use the parameters to initialize the object.
3.Copy Constructor:
A copy constructor is a member function which initializes an object using another object of the same class.